requirements.txt
msal>=1.20.0
requests>=2.28.1
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
python-dotenv>=0.21.0
rich>=12.6.0
PyYAML>=6.0
pytest>=7.2.0
black>=22.10.0
flake8>=6.0.0